[manuf] refactor manuf tests/lib organization
Manufacturing tests and provisioning libraries we mangled in one
directory. This separates the two so it is clear what is lib code vs
test code, and so it is clear what can be used to develop provisioning
binaries and future tests.
