--- sw/device/silicon_creator/manuf/lib/sram_device_info_flash_wr_functest.h
@@ -1,15 +1,0 @@
-// Copyright lowRISC contributors.
-// Licensed under the Apache License, Version 2.0, see LICENSE for details.
-// SPDX-License-Identifier: Apache-2.0
-
-#ifndef OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_LIB_SRAM_DEVICE_INFO_FLASH_WR_FUNCTEST_H_
-#define OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_LIB_SRAM_DEVICE_INFO_FLASH_WR_FUNCTEST_H_
-
-#include <stdint.h>
-
-/**
- * Type alias for the SRAM program entrypoint.
- */
-typedef bool sram_main(void);
-
-#endif  // OPENTITAN_SW_DEVICE_SILICON_CREATOR_MANUF_LIB_SRAM_DEVICE_INFO_FLASH_WR_FUNCTEST_H_
