+++ sw/device/silicon_creator/manuf/tests/empty_functest.c
