+++ sw/device/silicon_creator/manuf/tests/flash_device_info_flash_wr_functest.c
@@ -11,7 +11,7 @@
 #include "sw/device/lib/testing/test_framework/check.h"
 #include "sw/device/lib/testing/test_framework/ottf_main.h"
 #include "sw/device/silicon_creator/manuf/lib/isolated_flash_partition.h"
-#include "sw/device/silicon_creator/manuf/lib/test_wafer_auth_secret.h"
+#include "sw/device/silicon_creator/manuf/tests/test_wafer_auth_secret.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
