+++ sw/device/silicon_creator/manuf/tests/idle_functest.c
