+++ sw/device/silicon_creator/manuf/tests/sram_device_info_flash_wr_functest.c
@@ -18,7 +18,7 @@
 #include "sw/device/silicon_creator/manuf/lib/individualize_preop.h"
 #include "sw/device/silicon_creator/manuf/lib/isolated_flash_partition.h"
 #include "sw/device/silicon_creator/manuf/lib/otp_img.h"
-#include "sw/device/silicon_creator/manuf/lib/test_wafer_auth_secret.h"
+#include "sw/device/silicon_creator/manuf/tests/test_wafer_auth_secret.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
 #include "otp_ctrl_regs.h"  // Generated.
