+++ sw/device/silicon_creator/manuf/tests/sram_empty_functest.c
