+++ sw/device/silicon_creator/manuf/tests/sram_exec_test.c
